requirements.txt
xlrd~=1.2.0
demjson~=2.2.4
pandas~=1.0.4
networkx~=2.4
numpy~=1.18.5
graphviz~=0.14
grakel~=0.1b7
matplotlib~=3.2.2
scikit-learn~=0.20.4
mysql-connector-python~=8.0.21
